requirements.txt
torch>=2.0
transformers>=4.36
peft>=0.7
accelerate>=0.25
bitsandbytes>=0.41
Pillow>=10.0
matplotlib>=3.7
numpy>=1.24
scipy>=1.10
scikit-learn>=1.3
tqdm
datasets
